updated jodit editor to version 3.1.95 and fixed jodit addon
