+++ www/install.sql
@@ -24,7 +24,7 @@ INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "
 INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (19,	'Главное меню админки',	'block',	'main_menu',	1,	4,	14,	'menu',	's:1:"2";',	NULL,	1380884810);
 INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (29,	'Главное меню сайта',	'block',	'top_menu',	'0',	5,	14,	'menu',	's:1:"4";',	'',	1481295097);
 INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (40,	'301 редирект',	'addon',	'before',	'0',	1,	25,	'redirect_301',	'/test|/kontakty',	'',	1384346865);
-INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (41,	'Текстовый редактор jWysiwyg',	'addon',	'after',	1,	6,	28,	'jwysiwyg',	'',	'',	1390770516);
+INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (41,	'Текстовый редактор jodit',	'addon',	'after',	1,	6,	28,	'jodit',	'',	'',	1537785328);
 INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (42,	'Форма обратной связи',	'plugin',	'feedback',	'0',	7,	24,	'feedback',	'YTo2OntzOjY6ImZpZWxkcyI7YTo0OntpOjA7YTo1OntzOjQ6InR5cGUiO3M6NDoidGV4dCI7czo1OiJsYWJlbCI7czoxNToi0JLQsNGI0LUg0LjQvNGPIjtzOjg6InJlcXVpcmVkIjtzOjE6IjAiO3M6MTA6InZhbGlkYXRpb24iO3M6OToibm90X2VtcHR5IjtzOjEzOiJlcnJvcl9tZXNzYWdlIjtzOjA6IiI7fWk6MTthOjU6e3M6NDoidHlwZSI7czo0OiJ0ZXh0IjtzOjU6ImxhYmVsIjtzOjIxOiLQktCw0Ygg0YLQtdC70LXRhNC+0L0iO3M6ODoicmVxdWlyZWQiO3M6MToiMSI7czoxMDoidmFsaWRhdGlvbiI7czo1OiJwaG9uZSI7czoxMzoiZXJyb3JfbWVzc2FnZSI7czo1NToi0JLQstC10LTQuNGC0LUg0LLQtdGA0L3Ri9C5INC90L7QvNC10YAg0YLQtdC70LXRhNC+0L3QsCI7fWk6MjthOjU6e3M6NDoidHlwZSI7czo4OiJ0ZXh0YXJlYSI7czo1OiJsYWJlbCI7czoyNzoi0JLQsNGI0LUg0YHQvtC+0LHRidC10L3QuNC1IjtzOjg6InJlcXVpcmVkIjtzOjE6IjEiO3M6MTA6InZhbGlkYXRpb24iO3M6OToibm90X2VtcHR5IjtzOjEzOiJlcnJvcl9tZXNzYWdlIjtzOjA6IiI7fWk6MzthOjI6e3M6NDoidHlwZSI7czo2OiJzdWJtaXQiO3M6NToibGFiZWwiO3M6MTg6ItCe0YLQv9GA0LDQstC40YLRjCI7fX1zOjc6Im1haWxfdG8iO3M6MTI6ImluZm9Ac2l0ZS5ydSI7czoxMjoibWFpbF9zdWJqZWN0IjtzOjU3OiLQodC+0L7QsdGJ0LXQvdC40LUg0YHQviDRgdGC0YDQsNC90LjRhtGLINCa0L7QvdGC0LDQutGC0YsiO3M6OToibWFpbF9mcm9tIjtzOjE1OiJub3JlcGx5QHNpdGUucnUiO3M6MTQ6Im1haWxfdGV4dF9mcm9tIjtzOjEwOiLQoNC+0LHQvtGCIjtzOjE3OiJtYWlsX3RleHRfc3VjY2VzcyI7czo0ODoi0JLQsNGI0LUg0YHQvtC+0LHRidC10L3QuNC1INC+0YLQv9GA0LDQstC70LXQvdC+Ijt9',	'',	1514029556);
 INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (45,	'Слайдер CUBE',	'block',	'slider',	'0',	7,	31,	'cube_slider',	'a:5:{s:12:"aspect_ratio";d:2.8999999999999999;s:5:"speed";i:1000;s:9:"frequency";i:5000;s:10:"has_shadow";b:0;s:6:"slides";a:3:{i:0;a:4:{s:5:"image";s:34:"/uploads/modules/cube_slider/1.jpg";s:4:"text";s:0:"";s:8:"btn_text";s:0:"";s:4:"href";s:0:"";}i:1;a:4:{s:5:"image";s:34:"/uploads/modules/cube_slider/2.jpg";s:4:"text";s:0:"";s:8:"btn_text";s:0:"";s:4:"href";s:0:"";}i:2;a:4:{s:5:"image";s:34:"/uploads/modules/cube_slider/3.jpg";s:4:"text";s:0:"";s:8:"btn_text";s:0:"";s:4:"href";s:0:"";}}}',	'',	1481295676);
 INSERT INTO "activated_modules" ("id", "name", "type", "position", "back_end", "sort", "module_id", "module_dir", "params", "sections", "date_edit") VALUES (46,	'Боковое меню',	'block',	'aside',	'0',	4,	14,	'menu',	's:1:"7";',	'',	1481295416);
@@ -75,7 +75,7 @@ INSERT INTO "installed_modules" ("id", "name", "type", "dir", "version", "date_a
 /old_url.html|/ne_url.html
 /catalog|/katalog
 /contacts.html|/kontakty.html');
-INSERT INTO "installed_modules" ("id", "name", "type", "dir", "version", "date_add", "description") VALUES (28, 'Текстовый редактор jWysiwyg',  'addon',    'jwysiwyg', '1.0',  1390770516, '');
+INSERT INTO "installed_modules" ("id", "name", "type", "dir", "version", "date_add", "description") VALUES (28, 'Текстовый редактор jodit',  'addon',    'jodit', '3.1.95',  1537785328, '');
 INSERT INTO "installed_modules" ("id", "name", "type", "dir", "version", "date_add", "description") VALUES (29, 'Страничная навигация', 'plugin',   'pagination',   '1.0',  1480488062, 'Кнопки для страничной навигации');
 INSERT INTO "installed_modules" ("id", "name", "type", "dir", "version", "date_add", "description") VALUES (30, 'Анонс новостей',   'block',    'news_announce',    '2.0',  1480488062, 'Анонс новостей');
 INSERT INTO "installed_modules" ("id", "name", "type", "dir", "version", "date_add", "description") VALUES (31, 'Слайдер CUBE', 'block',    'cube_slider',  '1.4',  1481290914, 'Слайдер в виде вращающегося куба.');
